pyproject.toml
[project]
name = "trajectory-bench"
version = "0.1.0"
description = "Benchmark for evaluating LLM response equivalence on solved agent trajectories"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
litellm = ["litellm>=1.50.0"]
